requirements.txt
# requirements.txt per il progetto APC Training LLM (Struttura a Fasi)

streamlit>=1.32.0,<2.0.0
google-generativeai>=0.5.0,<1.0.0
faiss-cpu>=1.7.0,<2.0.0
# faiss-gpu # Alternativa se si usa GPU
numpy>=1.20.0,<2.0.0
pandas>=1.0.0,<3.0.0
protobuf

